Builk getting closer to completing.

+++ include/arch/pic30/thread.h
@@ -38,7 +38,7 @@ struct _callee_saved {
 typedef struct _callee_saved _callee_saved_t;
 
 struct _thread_arch {
-	uint32_t swap_return_value; /* Return value of z_swap() */
+	uint16_t swap_return_value; /* Return value of z_swap() */
 };
 
 typedef struct _thread_arch _thread_arch_t;
